- Several crash fixes - Array-based storage will now initialize containers to 0. - Fixed EOTS showing the wrong victor (visual only) - Logic error in anti-charm code - Relocated dismount vehicle code to a more appropriate location - Fixed a very low risk feeefeee in auras. - Gave CombatStatusHandler a memory barrier as a temporary measure to protect itself from annoying access violations. (todo: someone fix this for real) - Fixed Honorable Kill requirement achievements. - Fixed feeefeee crash with vehicles
git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/Hearthstone@872 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/SpellEffects.cpp
@@ -4912,7 +4912,7 @@ void Spell::SummonTotem(uint32 i) // Summon Totem
 		p_caster->m_SummonSlots[slot] = pTotem;
 
 	pTotem->SetTotemOwner(p_caster);
-	pTotem->SetTotemSlot(slot);
+	pTotem->SetTotemSlot( int32(slot) );
 
 	float landh;
 	if (p_caster->GetMapMgr() && p_caster->GetMapMgr()->IsCollisionEnabled())
